Visualitzar Detalls de ruta en api finalitzat, més endavant si puc ho milloraré

+++ BackEnd/src/main/java/cat/copernic/ymelero/EntreBicis/entity/Recompensa.java
@@ -33,7 +33,7 @@ public class Recompensa {
 
     private String observacions;
 
-    private Integer punts;
+    private Double punts;
 
     @Enumerated(EnumType.STRING)
     private EstatRecompensa estat;
